includes: start tidying up the supplied includes and library code

+++ include/stdint.h
@@ -19,15 +19,15 @@ typedef int16_t int_fast16_t;
 typedef uint32_t uint_fast32_t;
 typedef int32_t int_fast32_t;
 
-typedef int_least8_t int8_t;
-typedef uint_least8_t uint8_t;
-typedef int_least16_t int16_t;
-typedef uint_least16_t uint16_t;
-typedef int_least32_t int32_t;
-typedef uint_least32_t uint32_t;
+typedef int8_t int_least8_t;
+typedef uint8_t uint_least8_t;
+typedef int16_t int_least16_t;
+typedef uint16_t uint_least16_t;
+typedef int32_t int_least32_t;
+typedef uint32_t uint_least32_t;
 
-typedef intmax_t int32_t;
-typedef uintmax_t uint32_t;
+typedef int32_t intmax_t;
+typedef uint32_t uintmax_t;
 
 #define INT8_MIN	(-128)
 #define INT8_MAX	127
